Move element order control to Output tab
The selection of element order is tied to the output format, so it makes
sense to move it to that tab.

This change also updates some of the GUI Device names to be more clear.

Other ToggleRadioGroups in this file have names that end in Group to
differentiate them from their associated buttons.  Also use linear
and quadratic to emphasize we are setting element order -- and we are
not ordering the elements.