Use SetValMapVec to make toggle group work right
ToggleGroups assume (by default) that the added buttons have value
0,1,2,3... in order.  I.e. the first button = 0, second button=1, etc.

This assumption can be over-ridden by building a vector that maps
the button order to the desired values.  This can allow a list to start
at a non-zero value, or to skip values, etc.

Since we defined the enum FEA_LINEAR=1 and FEA_QUADRATIC=2, we need
to use SetValMapVec to make the toggle button setup work right.