Initialize and update element order flag
The call to .Init gives the IntParm a text name, a group, identifies
the ParmContainer that holds it (this), sets the default value (quadratic),
and sets the minimum and maximum values.

The update call makes sure the value stored is reflected in the GUI and
when the GUIDevice is clicked, the stored value is changed.